# apps/appointment-api/src/swagger/generated-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/Auth/signup": {
      "post": {
        "operationId": "AuthController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewUser" }
            }
          }
        },
        "responses": { "204": { "description": "User created successfully" } },
        "tags": ["Auth"]
      }
    },
    "/Auth/signin": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthUser" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return a list of Appointment",
            "content": {
              "application/json": { "schema": { "type": "string" } }
            }
          }
        },
        "tags": ["Auth"]
      }
    }
  },
  "info": {
    "title": "APPOINTMENT API",
    "description": "The appointment API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "NewUser": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "email": { "type": "string" },
          "role": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["ADMIN", "OPERATOR", "USER", "DISABLED"]
            }
          }
        },
        "required": ["username", "password", "email", "role"]
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": ["username", "password"]
      }
    }
  }
}